skip AsyncSocketTest.ConnectTimeout if the address is unreachable
Summary:
The code in AsyncSocketTest.ConnectTimeout attempts to pick an IP address to
connect to that will time out rather than failing immediately.  It checks to
see if IPv4 or IPv6 is available, but it cannot easily check if routes are
available.  If IPv6 is available but no routes are configured it will use IPv6,
and then the connect attempt will fail immediately rather than timing out.

This changes the test to simply report itself as skipped in this case rather
than failing.

Reviewed By: yfeldblum

Differential Revision: D6843332

fbshipit-source-id: 02ef3dd42f1d4b6e35edfe93b2e5e42c6666fa52