make the SKIP() macro report tests as successful
Summary:
The googletest framework does not have a built-in mechanism for skipping tests
at run-time.  Folly uses a `SKIP()` macro for some tests to report themselves
as skipped.  For our internal builds these are reported as failures with a
special message that gets handled by our test runner framework.  However for
open source builds it is better to report these tests as passing rather than as
failing.

Reviewed By: yfeldblum

Differential Revision: D6843331

fbshipit-source-id: f74f29354305703448e5757ddc0ec3e72380a8f7