update ThreadPoolExecutorTest to never try lowering niceness
Summary:
The code in PriorityThreadFactoryTest.ThreadPriority attempted to set the
thread niceness to 1.  This would fail if the process was already running at a
higher niceness value and was not being run as root.

This updates the test to get the current niceness first, and to use that plus 1
as the niceness value to change to.

Reviewed By: yfeldblum

Differential Revision: D6817548

fbshipit-source-id: 80e740f1fbe751865a418a58081bde4086b21679