Fix for deferredexecutor behaviour.
Summary: DeferredExecutor suffered from a case where boosting would not work if no work was present. This change modifies is so that on a call to via the next executor will be set in the DeferredExecutor. When add is called the callback will be stored. Whichever happens second actually enqueues the work on the chained executor. get uses via to correctly implement this.

Reviewed By: andriigrynenko

Differential Revision: D6663244

fbshipit-source-id: 74600d511373e0f6f7ca4128f25f3947a4981200