update the README to document building with cmake
Summary:
Update the build steps in the README to build using cmake instead of autoconf
on Linux platforms.

I haven't updated the instructions for Mac OS X or the
`bootstrap-osx-homebrew.sh` script primarily because I have not tested the
steps on Mac.  It sounds like the build is close to working on Mac, but
currently fails due to missing compiler flags to disable some warnings on
clang.

Reviewed By: akushner

Differential Revision: D6780814

fbshipit-source-id: 9c9c211c306d5b35235549ae451bb620098f37e3