FunctionSchedulerTest: honor timeFactor correctly
Summary:
Most of the tests in FunctionSchedulerTest sleep for lengths of time based on
the timeFactor constant set at the top of the file.  However a couple of places
that need to sleep for non-integral timeFactor durations used manual `usleep()`
calls instead.  This made the tests break if you change the `timeFactor` value.

This changes `delay()` to accept floating point numbers, and updates the tests
that were using hard-coded `usleep()` calls to use `delay()` instead.

Reviewed By: yfeldblum

Differential Revision: D6880258

fbshipit-source-id: 0cfdf84d3ba27aa2acd1f5210becbc7306b857bf