FunctionSchedulerTest: increase timeFactor to 400ms
Summary:
Many of the tests in FunctionSchedulerTest appear to frequently fail on
Travis CI builds.  It looks like this is probably due to timing issues where
events do not fire within the expected time interval.

This diff increases the `timeFactor` value to hopefully make the tests more
reliable, at the expense of making them take longer to run.

It would be nicer if we had a better solution here that did not involve
potentially unreliable timing-dependent checks.  Increasing `timeFactor` seems
like the easiest solution in the short-term though, and avoids just disabling
the tests entirely.

Reviewed By: yfeldblum

Differential Revision: D6880256

fbshipit-source-id: 59da3b20f2bb9c982834752942a984e3f91442fd