enable FOLLY_DISABLE_UNDEFINED_BEHAVIOR_SANITIZER macro in ASAN
Summary:
It's common to use UBSAN with ASAN at the same time, and we'll merge them eventually too. As a result we can check for ASAN defined macros to conditionally compile code for UBSAN.

We still have a few build modes that use standalone UBSAN, and unlike other sanitizers, UBSAN doesn't define a magic macro that we could use. In the past we added UNDEFINED_SANITIZER to our compiler flags, but I'd like to go away from it towards a more standard solution (eg check for ASAN).

Reviewed By: yfeldblum, luciang

Differential Revision: D6931054

fbshipit-source-id: 4b298d1cd72cecd693bf99eda309568ef019c544