Gate hazptr static_assert which is guaranteed to fail in MSVC
Summary:
std::atomic is no longer trivial, and VS2015 and VS2017 both
fail this static assert, as will future versions of conforming compilers.
Updates to folly mean that the hazptr header is transitively included in
more places which makes it easier to trip this even when not using hazptr.
Disabling this assert to visual studio at least saves users a headache.

Reviewed By: yfeldblum, magedm

Differential Revision: D7006587

fbshipit-source-id: 8481e1d5668aee8cab703a04b2a390550c210a2c