[CMake] Closer to compiling with Clang
Summary:
This brings us closer to being able to compiling the open
source version of Folly with Clang. There are a few other
open pull requests that need to be addressed to allow us to
compile on Clang without any warnings that aren't already
explicitly disabled. See pull requests `745` and `756`.