Closer to compiling with Clang
Summary:
This brings us closer to being able to compile the open
source version of Folly with Clang. There are a few other
open pull requests that need to be addressed to allow us to
compile on Clang without any warnings that aren't already
explicitly disabled. See pull requests #745 and #756.

Closes https://github.com/facebook/folly/pull/772

Reviewed By: simpkins

Differential Revision: D7021595

Pulled By: yfeldblum

fbshipit-source-id: 350b972c74502ca2bc147f21428c622b19805b89