Fix master branch broken for Clang
Summary:
Clang does not support the warning option "-Wclobbered", so we cannot
do a pragma push and pop for this warning. So, we push and pop this
particular warning for GCC only.

As it stands currently, Clang accepts `#pragma GCC` directives. If this
sort of thing comes up more in the future, it may make sense for Folly
to define a macro for easily disabling warnings for either just GCC or
just Clang.

simpkins yfeldblum
Closes https://github.com/facebook/folly/pull/779

Reviewed By: simpkins

Differential Revision: D7059916

Pulled By: yfeldblum

fbshipit-source-id: a6d420dc6ae435e3be1229b2bb421910bc8794ff