Fix DynamicConverter to work with vector<bool>
Summary:
Introduce a specialization for `std::vector<bool>` as the existing
specialization for "other ranges" is not sufficient for binding the
proxy iterator (`bit_const_reference`) to an lvalue when iterating over
the `vector<bool>`.

I do not think you can simply change the type from `const auto&` to `auto&&` when iterating over the container for the "other ranges" specialization.  This is because we cannot bind an rvalue to an lvalue reference (which is the function argument type for our `construct` specializations).

See #752 and https://github.com/facebook/folly/pull/756 for more info.
Closes https://github.com/facebook/folly/pull/781

Reviewed By: andrewjcg

Differential Revision: D7078659

Pulled By: yfeldblum

fbshipit-source-id: 0089183a3c68d20ed14e483859e950dfb4ae729e