logging: add an FB_LOG_RAW() macro
Summary:
`FB_LOG_RAW()` is similar to `FB_LOG()` but allows the caller to pass in the
log level as a variable, and to specify an alternate file name and line number.

This is useful for callers that wish to implement their own logging utility
functions that log a message, but pass through their caller's file name and
line number.

Reviewed By: wez

Differential Revision: D7128091

fbshipit-source-id: 1fec158a9312b9c9dc53cf3dc87004d42f9806c6