Ensure that SemiFuture executor is correctly chained to deal with deferred work.
Summary: There was previously a bug in that defer used within a continuation would not correctly chain the deferred executor on the executor in the returned (and hence in the original) folly::Future like it would when .via was called on a SemiFuture. This fixes that situation by returning a semifuture from a continuation by correctly chaining it onto the outer future's executor.

Reviewed By: yfeldblum

Differential Revision: D7156241

fbshipit-source-id: ed6db3ad0de7a921b72459fcd2b1228114832853