Remove Promise::getFuture from Future-inl.h 4/4 - window
Summary: Modify folly::window to replace Promise::getFuture calls with Promise::getSemiFuture. The requirement for an executor on the semifuture here meant that I had to modify all of the SemiFutures with an executor in the vector. This may come with a small performance hit in some cases, but appears to be behaviourally better.

Reviewed By: yfeldblum

Differential Revision: D7104920

fbshipit-source-id: 7d105df553c28bb362b37ff64270ee16bf59fbd6