Add a buckets + size seqlock
Summary:
There is currently an unnecessary data dependency loading the
current backing buckets + size, reading the size, then again indexing
in to the array.

Rework the Buckets struct to make a single allocation, and store the bucket_count_ in the Segment directly.  The bucket pointer and segment are protected by a seqlock.

Reviewed By: ot

Differential Revision: D7167919

fbshipit-source-id: 10ddceffad19f54c790b3ab7a87260445571f81d