logging: add a getBaseLoggingConfig() function
Summary:
Update folly::initLogging() to call getBaseLoggingConfig() to get a base
configuration string that is applied before the argument that it was called
with.  getBaseLoggingConfig() is defined as a weak symbol, and is intended to
be overridden on a per-executable basis if desired.

This allows individual programs to control their default logging settings, but
still let them be overridden via a command line flag.  The command line flag is
used to update this base configuration.  The command line flag can fully
override all of the base configuration settings if desired, but settings from
the base config that are not overridden will still be used.

For example, if the base configuration defines settings for the categories
"foo" and "bar", and the command line flag defines settings for the categories
"bar" and "wub", the base configuration settings for "foo" will still be used.

Reviewed By: yfeldblum

Differential Revision: D7164832

fbshipit-source-id: e172f746d7bd004948872adbbb87c597765e283c