Temporarily remove onError returning a future that completes on the original executor due to production issues.
Summary: https://fb.facebook.com/groups/fbcode.investigations/permalink/407299856363159/ showed issues coming from this diff. The change to return an executor-tied future from onError was intentional, and correct behaviour, but it was a change that seems to have caused issues. Temporarily removing it to rethink how we should approach it.

Reviewed By: andriigrynenko

Differential Revision: D7219231

fbshipit-source-id: 95b91f3a5ea7d854d5df052e6cd5851c7e6fbdab