Allowing folly::dynamics of object/array type to be hashed
Summary:
Currently, an error is thrown if a user tries to hash an instance of `folly::dynamic` which is of type `object` or `array`. Primitive underlying types are fine.

There doesn't seem to be a strict reason for this, as objects and arrays are just as hashable as primitive types. Additionally, the inability to hash here causes some unnecessary drawbacks (e.g. interacting with `unordered_map`s).

This diff uses the existing dependency on `folly/hash/Hash.h` to start allowing `folly::dynamic` objects/arrays to be hashed as well.

Reviewed By: ot

Differential Revision: D7219760

fbshipit-source-id: 1dff3f99bdacb66f719c614040cbd668ce51a33e