Check that fiber suspends are not occuring in exception catch blocks
Summary:
Fiber suspends are incompatible with C++ exception handling. We should enforce that callers are not suspending fibers while handling exceptions.
See https://fb.facebook.com/groups/560979627394613/permalink/1119586164867287/

NOTE: This check currently fires not just for baton.wait() (and all things based on baton), but also for calling runInMainContext. I'm unsure of the safety of this.

Reviewed By: yfeldblum

Differential Revision: D6695110

fbshipit-source-id: cf36edbab9f0573cbf639bf79d0c267681ac5914