save 8 bytes and a bit of work in F14Vector{Map,Set}
Summary:
F14VectorMap and F14VectorSet use ItemIter to interact with F14Table,
but they don't actually need to be able to identify the begin ItemIter.
This diff removes both the storage for packedBegin_ and the code to
update it (present in all insert and erase functions) for that policy.
This is a small CPU optimization, and reduces sizeof(F14Vector{Map,Set})
from 32 bytes to 24.

Reviewed By: shixiao

Differential Revision: D7181769

fbshipit-source-id: 318c63584543fca339a002d8bcb5872a7bd0bdd3