Add an option to the MSVC build to suppress boost's out of date config messages
Summary:
MSVC moves faster than boost updates, so it's not unusual to be building with an MSVC newer than the boost being used was able to be tested against. This causes Boost to emit a message at compile time for every compilation unit it is included in.
Luckly, boost provides a way out of the message, so add a config option to do so and enable it by default.

Reviewed By: yfeldblum

Differential Revision: D7388583

fbshipit-source-id: 8ecddc048df432df5200f4601e4121bdbd275220