Fix the forward declarations of F14 under MSVC
Summary:
Due to a quirk in how MSVC does name resolution, if a default value is declared in the forward declaration of a template, and the definition is made available in the same unit, the default values in the forward declaration must use the names given in the definition, not the declaration.
To keep it sane, just use the same template parameter names for both the declaration and definition.

(Note: this ignores all push blocking failures!)

Reviewed By: shixiao

Differential Revision: D7405057

fbshipit-source-id: dda14043fec30fa505bb0bb3e2e31d0abb52ba1f