folly/Singleton: remove use of `link_whole`
Summary:
We used `link_whole` to make sure the `setStackTraceGetter` constructor is
always linked when using Singleton's to setup symbolizer for stack traces.
However, using `link_whole` means we always link in the singleton library, even
if it's never used.

This diff merges `SingletonStackTrace.cpp` into `Singleton.cpp` to avoid the
need for `link_whole` and to link the former if and only if we use the latter.

Reviewed By: pixelb

Differential Revision: D7230733

fbshipit-source-id: de9be51a4eefee201563523d25800ab19ea777e3