Reduce recursion, for ASAN at -O0
Summary:
At `-O0` each coroutine call results in a heap allocation. This
combined with the increased memory pressure when using ASAN
results in an invalid memory access beyond the maximum address.
Avoid the problem by limiting the recursion to 5,000 frames deep,
instead of 10,000. This still fulfills the intent of the test,
but without pushing the limits of ASAN.

Reviewed By: wqfish

Differential Revision: D7503976

fbshipit-source-id: 7850a9d9db23328d42c25b74a1290caa7773e6a7