document transferrability of F14HashToken between instances
Summary:
F14HashToken encapsulates the work of hash_function(), which
is the same across F14 containers that have the same key_type and equal
hash_function() (same type and operator==).  This diff adds documentation
that callers may rely on this fact, which means they can use this feature
to reduce work if a single key is access in multiple maps/sets.

(Note: this ignores all push blocking failures!)

Reviewed By: shixiao

Differential Revision: D7506737

fbshipit-source-id: 321032d619434501b1e0544962e54732bfc4c970