Make hazptr_obj copyable and moveable
Summary: hazptr_obj is copyable and moveable, but doing so would cause a bug, because the next ptr would point to the object that was copied/moved from, which can oftentimes be garbage or another object that has not yet retired. This diff changes the copy/move constructors and assignment operators to simply check that the rhs has not been retired.

Reviewed By: magedm

Differential Revision: D7539444

fbshipit-source-id: f9325c7b886fcf4b553344ee37f2355ca41268b6